JobPortal/requirements.txt
asgiref==3.9.1
Django==5.2.4
django-ckeditor==6.7.3
django-js-asset==3.1.2
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
sqlparse==0.5.3
tzdata==2025.2
